refactor: MainViewController에서 NumberFormatter 기능을 extention을 통해 분리
